More moving of implementation to .cpp files
