+++ src/komodo.h
@@ -42,7 +42,7 @@ bool check_pprevnotarizedht();
 #include "komodo_utils.h"
 #include "komodo_curve25519.h"
 
-#include "komodo_cJSON.c"
+#include "komodo_cJSON.h"
 #include "komodo_bitcoind.h"
 #include "komodo_interest.h"
 #include "komodo_pax.h"
